requirements.txt
pandas
xlsxwriter
matplotlib
googleapiclient
imageio
google-cloud-language
xlrd
